Restructure logical_test
